Remove the deprecated APIs in JS envs.
In the process, the implementation of
`org.scalajs.jsenv.nodejs.JSDOMNodeJSEnv` is transferred to the
new `org.scalajs.jsenv.jsdomnodejs.JSDOMNodeJSEnv`.
